ログストアをログコンポーネントに移動
